shuffleArray moved into shared/helpers
