+++ src/modules/Player/components/Player.vue
